also move ceiling_set_info
